LayoutTests/imported/w3c/web-platform-tests/svg/types/scripted/SVGGraphicsElement.getBBox-10-expected.txt

99
There are issues of dynamic loading required for tiling. According to 'postpone' attribute of Resource Priorities, the dynamic loading is controlled by positional relation with bounding box of embedded contents and container's viewport. However, each bounding boxes of embedded contents should be whole earth basically when this method is used. (green part on Example) Tiling is impossible unless this situation is changed.

PASS text with zoom
PASS image with zoom
PASS path with zoom
FAIL foreignObject with zoom assert_approx_equals: fo1.getBBox().x expected 2 +/- 0.1 but got 0
